ProyectoSena/requirements.txt
Django==2.0.4
django-allauth==0.36.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
et-xmlfile==1.0.1
gunicorn==19.4.5
openpyxl==2.5.3
Pillow==5.1.0
django-heroku==0.3.1
dj-database-url==0.5.0
whitenoise==3.3.1